+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Enhancements
 
 * **AWS IoT**
-  * AWSIotMqttManager now exposes an API `setUsernameFields` for clients to pass additional key-value pairs which are appended to the username used for connection.
+  * AWSIotMqttManager now exposes an API `addUserMetaData` for clients to pass additional key-value pairs which are appended to the username used for connection.
 
 ## [Release 2.12.3](https://github.com/aws/aws-sdk-android/releases/tag/release_v2.12.3)